+++ test/lib/trace.spec.js
@@ -524,6 +524,156 @@ describe("lib/trace", () => {
         "node_modules/two/package.json"
       ]));
     });
+
+    it("handles try/catch missing requires", async () => {
+      mock({
+        "hi.js": `
+          require("one");
+
+          const { aFunction } = require("nested-first-level");
+          const { aFile } = require("nested-trycatch-requireresolve");
+        `,
+        node_modules: {
+          one: {
+            "package.json": stringify({
+              main: "index.js"
+            }),
+            "index.js": `
+              module.exports = {
+                one: () => "one",
+                two: () => require("two").two
+              };
+            `
+          },
+          two: {
+            "package.json": stringify({
+              main: "index.js"
+            }),
+            "index.js": `
+              module.exports = {
+                two: () => "two"
+              };
+            `
+          },
+          "nested-first-level": {
+            "package.json": stringify({
+              main: "lib/index.js"
+            }),
+            lib: {
+              "index.js": `
+                const { aFunction } = require("nested-trycatch-require");
+
+                module.exports = {
+                  aFunction
+                };
+              `
+            }
+          },
+          "nested-trycatch-require": {
+            "package.json": stringify({
+              main: "index.js"
+            }),
+            "index.js": `
+              let aFunction;
+              try {
+                aFunction = () => import("doesnt-exist/with/path.js");
+              } catch (err) {
+                aFunction = () => null;
+              }
+
+              module.exports = { aFunction };
+            `
+          },
+          "nested-trycatch-requireresolve": {
+            "package.json": stringify({
+              main: "index.js"
+            }),
+            "index.js": `
+              let noFile = null;
+              try {
+                noFile = require.resolve("also-doesnt-exist");
+              } catch (err) {}
+
+              module.exports = { noFile };
+            `
+          }
+        }
+      });
+
+      expect(await traceFile({
+        srcPath: "hi.js",
+        allowMissing: {
+          "nested-trycatch-require": [
+            "doesnt-exist"
+          ],
+          "nested-trycatch-requireresolve": [
+            "also-doesnt-exist"
+          ]
+        }
+      })).to.eql(fullPath([
+        "node_modules/nested-first-level/lib/index.js",
+        "node_modules/nested-first-level/package.json",
+        "node_modules/nested-trycatch-require/index.js",
+        "node_modules/nested-trycatch-require/package.json",
+        "node_modules/nested-trycatch-requireresolve/index.js",
+        "node_modules/nested-trycatch-requireresolve/package.json",
+        "node_modules/one/index.js",
+        "node_modules/one/package.json",
+        "node_modules/two/index.js",
+        "node_modules/two/package.json"
+      ]));
+    });
+
+    it("still errors on missing imports in a catch", async () => {
+      mock({
+        "hi.js": `
+          const { aFunction } = require("nested-first-level");
+        `,
+        node_modules: {
+          "nested-first-level": {
+            "package.json": stringify({
+              main: "lib/index.js"
+            }),
+            lib: {
+              "index.js": `
+                const { aFunction } = require("nested-trycatch-require");
+
+                module.exports = {
+                  aFunction
+                };
+              `
+            }
+          },
+          "nested-trycatch-require": {
+            "package.json": stringify({
+              main: "index.js"
+            }),
+            "index.js": `
+              let aFunction;
+              try {
+                aFunction = () => import("doesnt-exist/with/path.js");
+              } catch (err) {
+                aFunction = () => null;
+              }
+
+              module.exports = { aFunction };
+            `
+          }
+        }
+      });
+
+      await expect(traceFile({
+        srcPath: "hi.js",
+        allowMissing: {
+          "nested-first-level": [
+            // This won't be a permitted missing because only `nested-trycatch-require` is checked.
+            "doesnt-exist"
+          ]
+        }
+      })).to.be.rejectedWith(
+        /Encountered resolution error in .*nested-trycatch-require.* for doesnt-exist.*/
+      );
+    });
   });
 
   describe("traceFiles", () => {