+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changes
 
 ## 0.2.2
 
+* Bug/Feature: Allow permissive handling of try/catch `require`s with `allowMissing` parameter.
+  [#19](https://github.com/FormidableLabs/trace-deps/issues/19)
 * Upgrade `node-acorn` to `^2.0.0`.
 
 ## 0.2.1